v2/app/requirements.txt
flask==3.1.2
Flask-login
Flask-sqlalchemy
Flask-Limiter
requests==2.32.5
python-dotenv
psycopg2
sqlalchemy
pandas
openpyxl